requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pydantic==2.10.3
pydantic_core==2.27.1
requests==2.32.4
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
kucoin-universal-sdk
python-dotenv
ta
pandas
rich
textual
websockets